requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
pytest==7.4.3
python-dotenv==1.0.0 